+++ lib/td_dd/search/index_worker.ex
@@ -7,7 +7,6 @@ defmodule TdDd.Search.IndexWorker do
 
   use GenServer
 
-  alias TdDd.DataStructures.PathCache
   alias TdDd.Search.Indexer
 
   require Logger
@@ -73,14 +72,12 @@ defmodule TdDd.Search.IndexWorker do
 
   @impl true
   def handle_cast({:reindex, :all}, state) do
-    PathCache.refresh()
     do_reindex(:all)
     {:noreply, state}
   end
 
   @impl true
   def handle_cast({:reindex, data_structure_ids}, state) do
-    PathCache.refresh()
     do_reindex(data_structure_ids)
     {:noreply, state}
   end